Add bind_rows()
Until now, bind_rows is the only function in quanteda.tidy that doesn't build upon a generic function in dplyr. So I defined bind_rows as a generic function first. With the .default method, bind_rows works as expected for dataframes. The .corpus method extends the functionality of bind_rows to objects with a corpus class. I thought it might be useful to allow the user to be able to specifcy the metadata of the combined corpora with a list object. If such object is not supplied in the .meta argument, bind_rows.corpora will use the metadata of the first corpus by default.